fix: resolve compilation errors and improve build configuration
- Add virtual destructors to FailHandler and ErrorHandler classes in Sender.h
- Add [[maybe_unused]] attributes to unused parameters across codebase
- Fix member initialization order in BarGenerator::ctor (utility.cpp)
- Fix sign comparison error in JsonFormat (utils.cpp)
- Remove invalid line continuation character in engine.h comment
- Refactor compiler warning configuration to use global options.cmake
- Handle third-party library warnings (-Wunused-variable, -Wnull-pointer-subtraction) globally instead of per-file
- Remove file-specific CMake warning suppression from core/api/CMakeLists.txt
- All build targets now compile successfully (trader, libapi.dylib, libengine.dylib, etc.)

+++ core/exchange/utils.cpp
@@ -23,7 +23,7 @@ AString getLevelStr(int level) {
 AString JsonFormat(AString jsonStr) {
     int level = 0;
     AString jsonForMatStr;
-    for(int i=0;i<jsonStr.length();i++){
+    for(size_t i=0;i<jsonStr.length();i++){
         char c = jsonStr.at(i);
         if(level>0&&'\n'==jsonForMatStr.at(jsonForMatStr.length()-1)){
             jsonForMatStr += getLevelStr(level);
